requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.7.0
psycopg2-binary>=2.9.9
